Replace switch with if on resources, which will be non-final in gradle 8

+++ app/src/main/java/se/lublin/mumla/servers/FavouriteServerAdapter.java
@@ -44,19 +44,18 @@ public class FavouriteServerAdapter extends ServerAdapter<Server> {
 
     @Override
     public boolean onPopupItemClick(Server server, MenuItem menuItem) {
-        switch (menuItem.getItemId()) {
-            case R.id.menu_server_edit:
-                mListener.editServer(server);
-                return true;
-            case R.id.menu_server_share:
-                mListener.shareServer(server);
-                return true;
-            case R.id.menu_server_delete:
-                mListener.deleteServer(server);
-                return true;
-            default:
-                return false;
+        int itemId = menuItem.getItemId();
+        if (itemId == R.id.menu_server_edit) {
+            mListener.editServer(server);
+            return true;
+        } else if (itemId == R.id.menu_server_share) {
+            mListener.shareServer(server);
+            return true;
+        } else if (itemId == R.id.menu_server_delete) {
+            mListener.deleteServer(server);
+            return true;
         }
+        return false;
     }
 
     public static interface FavouriteServerAdapterMenuListener {
